Практическое задание 5: Модуляция
